dialog を components に移動
