+++ web/src/components/course/EditableCoursesTable/index.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from "react";
 import courseApi from "../../../api/course";
 import type { Course, Day, UserID } from "../../../common/types";
 import FullScreenCircularProgress from "../../common/FullScreenCircularProgress";
-import SelectCourseDialog from "../SelectCourseDialog";
 import CoursesTableCore from "../components/CoursesTableCore";
+import SelectCourseDialog from "../components/SelectCourseDialog";
 
 type Props = {
   userId: UserID;
