+++ web/src/components/course/components/SelectCourseDialog.tsx
@@ -15,9 +15,9 @@ import {
 } from "@mui/material";
 import IconButton from "@mui/material/IconButton";
 import { useEffect, useState } from "react";
-import courseApi from "../../api/course";
-import { DAY_TO_JAPANESE_MAP } from "../../common/consts";
-import type { Course, Day } from "../../common/types";
+import courseApi from "../../../api/course";
+import { DAY_TO_JAPANESE_MAP } from "../../../common/consts";
+import type { Course, Day } from "../../../common/types";
 import CourseDeleteConfirmDialog from "./CourseDeleteConfirmDialog";
 import CourseRegisterConfirmDialog from "./CourseRegisterConfirmDialog";
